fix: moved initial data fetching to header container
