+++ src/containers/home/Home.js
@@ -1,19 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import './Home.css';
-import { fetchLocationAction, fetchCategoriesAction, fetchSkillsActionCreator } from '../../store/actions/actions';
 import { BulmaBoiler, TagCloud } from '../../components/'
 
 class Home extends Component {
-  
-  componentWillMount() {
-    if (this.props.location.status === 'unloaded') this.props.getLocation();
-  }
-
-  componentDidUpdate() {
-    if (this.props.location.status === 200 && this.props.categories.status === 'unloaded') this.props.getCategories();
-    if (this.props.location.status === 200 && this.props.skills.status === 'unloaded') this.props.getSkills();
-  }
 
   render () {
     return (
@@ -34,17 +24,4 @@ const mapStateToProps = (state) => ({
   skills: state.skills
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  getLocation: () => dispatch(fetchLocationAction),
-  getCategories: () => dispatch(fetchCategoriesAction),
-  getSkills: (location) => dispatch(fetchSkillsActionCreator(location)),
-});
-
-const mergeProps = (stateProps, dispatchProps, ownProps) => ({
-  ...ownProps,
-  ...stateProps,
-  ...dispatchProps,
-  getSkills: () => dispatchProps.getSkills(stateProps.location)
-});
-
-export default connect(mapStateToProps, mapDispatchToProps, mergeProps)(Home);
+export default connect(mapStateToProps)(Home);
